move decryptResult into result-utils
