+++ src/cli/cmd/iexec-result.js
@@ -7,7 +7,7 @@ import fsExtra from 'fs-extra';
 import { Buffer } from 'buffer';
 import { join } from 'path';
 import { generateKeyPair } from 'crypto';
-import { decryptResult } from '../../common/utils/utils.js';
+import { decryptResult } from '../../common/utils/result-utils.js';
 import { getResultEncryptionKeyName } from '../../common/utils/secrets-utils.js';
 import { checkWeb2SecretExists } from '../../common/sms/check.js';
 import { pushWeb2Secret } from '../../common/sms/push.js';
