+++ src/lib/utils.js
@@ -7,7 +7,7 @@ export {
   encodeTag,
   decodeTag,
   sumTags,
-  decryptResult,
 } from '../common/utils/utils.js';
+export { decryptResult } from '../common/utils/result-utils.js';
 export { NULL_ADDRESS, NULL_BYTES32 } from '../common/utils/constant.js';
 export { getSignerFromPrivateKey } from '../common/utils/signers.js';
